pyproject.toml
[tool.poetry]
name = "hw17"
version = "0.1.1"
description = "Home Work 17"
authors = ["RusLAN Polikarpov <ruslan.polikarpov.2014@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0"
flask = "^2.3.2"
jwt = "^1.3.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
